Move validator endpoints to a separate module.
